+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/WithRequestCallable.java
@@ -31,8 +31,7 @@ public class WithRequestCallable<C extends Client> implements FunctionCallable<C
 
   @Override
   public <O> O call(Function<C, O> function) {
-    try (C newClient = (C) this.client.newClient(requestConfig).adapt(this.client.getClass())) {
-      return function.apply(newClient);
-    }
+    C newClient = (C) this.client.newClient(requestConfig).adapt(this.client.getClass());
+    return function.apply(newClient);
   }
 }