+++ openshift-client-api/src/main/java/io/fabric8/openshift/client/DefaultOpenShiftClient.java
@@ -20,6 +20,7 @@ import io.fabric8.kubernetes.client.KubernetesClientBuilder;
 import io.fabric8.kubernetes.client.http.HttpClient;
 import io.fabric8.kubernetes.client.http.HttpClient.Builder;
 import io.fabric8.kubernetes.client.http.HttpClient.Factory;
+import io.fabric8.kubernetes.client.http.StandardHttpClientBuilder;
 import io.fabric8.kubernetes.client.utils.HttpClientUtils;
 import io.fabric8.kubernetes.client.utils.Serialization;
 
@@ -68,14 +69,27 @@ public class DefaultOpenShiftClient extends NamespacedOpenShiftClientAdapter {
 
         @Override
         public Builder newBuilder() {
-          // should not be called
           throw new UnsupportedOperationException();
         }
 
         @Override
-        public HttpClient createHttpClient(Config config) {
-          return httpClient;
+        public Builder newBuilder(Config config) {
+          return new StandardHttpClientBuilder<HttpClient, HttpClient.Factory, StandardHttpClientBuilder<HttpClient, HttpClient.Factory, ?>>(
+              null) {
+
+            @Override
+            public HttpClient build() {
+              return httpClient;
+            }
+
+            @Override
+            protected StandardHttpClientBuilder<HttpClient, HttpClient.Factory, StandardHttpClientBuilder<HttpClient, HttpClient.Factory, ?>> newInstance(
+                Factory clientFactory) {
+              return null;
+            }
+          };
         }
+
       });
     }
     this.init(builder.build());